refactor: move elderly school libs to lib/elderly-school/
- lib/elderlySchoolDashboard.js → lib/elderly-school/dashboard.js
- lib/elderlySchoolImport.js → lib/elderly-school/import.js
- lib/elderlyMentalHealth.js → lib/elderly-school/mentalHealth.js
